mpi adaptor header
