+++ include/boost/multi/adaptors/mpi/test/mpi.cpp
@@ -4,231 +4,15 @@
 
 #include <boost/multi/array.hpp>
 
+#include <boost/multi/adaptors/mpi.hpp>
+
 #define OMPI_SKIP_MPICXX 1
 #include <mpi.h>
 
 #include <boost/core/lightweight_test.hpp>
 
-#include <cassert>   // for assert
 #include <iostream>  // for std::cout
-#include <limits>    // for numeric_limits  NOLINT(misc-include-cleaner)
-#include <utility>   // for exchange, move
 #include <vector>
-
-namespace boost::multi::mpi {
-
-class data {
-	void*        buf_;
-	MPI_Datatype datatype_;
-
- public:
-	template<class It>
-	explicit data(It first)                                            // NOLINT(cppcoreguidelines-pro-type-member-init,hicpp-member-init)
-	: buf_{const_cast<void*>(static_cast<void const*>(first.base()))}  // NOLINT(cppcoreguidelines-pro-type-const-cast)
-	{
-		MPI_Type_vector(
-			1, 1,
-			first.stride(),
-			MPI_INT, &datatype_
-		);
-
-		MPI_Type_commit(&datatype_);  // type cannot be used until committed, in communication operations at least
-	}
-
-	data(data const&) = delete;
-	data(data&&)      = delete;
-
-	auto operator=(data const&) = delete;
-	auto operator=(data&&)      = delete;
-
-	~data() { MPI_Type_free(&datatype_); }
-
-	auto buffer() const { return buf_; }
-	auto type() const { return datatype_; }
-};
-
-template<class Layout>
-auto create_subarray_aux(
-	Layout        lyt,
-	int           subcount,
-	MPI_Datatype  old_datatype,
-	MPI_Datatype* new_datatype
-) -> int {
-	MPI_Datatype sub_type;  // NOLINT(cppcoreguidelines-init-variables)
-
-	if constexpr(Layout::dimensionality == 1) {
-		MPI_Type_dup(old_datatype, &sub_type);
-	} else {
-		create_subarray_aux(lyt.sub(), lyt.sub().size(), old_datatype, &sub_type);
-	}
-
-	int dt_size;  // NOLINT(cppcoreguidelines-init-variables)
-	MPI_Type_size(old_datatype, &dt_size);
-
-	assert(lyt.stride() * dt_size <= std::numeric_limits<MPI_Aint>::max());
-	{
-		MPI_Datatype vector_datatype;  // NOLINT(cppcoreguidelines-init-variables)
-		MPI_Type_create_hvector(
-			subcount, 1,
-			lyt.stride() * dt_size,
-			sub_type, &vector_datatype
-		);
-
-		MPI_Type_create_resized(vector_datatype, 0, lyt.stride() * dt_size, new_datatype);
-		MPI_Type_free(&vector_datatype);
-	}
-	MPI_Type_free(&sub_type);
-	return MPI_SUCCESS;
-}
-
-template<class Size = int>
-class skeleton {
-	Size         count_;
-	MPI_Datatype datatype_;
-
-	skeleton() : datatype_{MPI_DATATYPE_NULL} {}
-
-	auto operator=(skeleton&& other) & noexcept -> skeleton& {
-		count_    = other.count_;
-		datatype_ = std::exchange(other.datatype_, MPI_DATATYPE_NULL);
-		return *this;
-	}
-
-	template<class Layout>
-	// NOLINTNEXTLINE(cppcoreguidelines-pro-type-member-init,hicpp-member-init,fuchsia-default-arguments-declarations)
-	skeleton(Layout const& lyt, MPI_Datatype dt, Size subcount) : count_{static_cast<Size>(lyt.size())} {
-		assert(lyt.size() <= std::numeric_limits<Size>::max());
-
-		MPI_Datatype              sub_type;  // NOLINT(cppcoreguidelines-init-variables)
-		[[maybe_unused]] skeleton sk;        // NOLINT(misc-const-correctness)
-		if constexpr(Layout::dimensionality == 1) {
-			sub_type = dt;
-		} else {
-			sk       = skeleton(lyt.sub(), dt, lyt.sub().size());
-			sub_type = sk.type();
-		}
-
-		int dt_size;  // NOLINT(cppcoreguidelines-init-variables)
-		MPI_Type_size(dt, &dt_size);
-
-		{
-			MPI_Datatype vector_datatype;  // NOLINT(cppcoreguidelines-init-variables)
-			MPI_Type_create_hvector(
-				subcount, 1,
-				lyt.stride() * dt_size,
-				sub_type, &vector_datatype
-			);
-
-			MPI_Type_create_resized(vector_datatype, 0, lyt.stride() * dt_size, &datatype_);
-			MPI_Type_free(&vector_datatype);
-		}
-	}
-
- public:
-	skeleton(skeleton&& other) noexcept
-	: count_{other.count_}, datatype_{std::exchange(other.datatype_, MPI_DATATYPE_NULL)} {}
-
-	template<class Layout>
-	skeleton(Layout const& lyt, MPI_Datatype dt)
-	: skeleton{lyt, dt, 1} {
-		MPI_Type_commit(&datatype_);
-	}
-
-	skeleton(skeleton const&) = delete;
-
-	auto operator=(skeleton const&) = delete;
-
-	~skeleton() {
-		if(datatype_ != MPI_DATATYPE_NULL) {
-			MPI_Type_free(&datatype_);
-		}
-	}
-
-	auto count() const { return count_; }
-	auto type() const& { return datatype_; }
-	auto type() && { return std::exchange(datatype_, MPI_DATATYPE_NULL); }
-};
-
-template<class Layout>
-auto create_subarray(Layout const& lyt, MPI_Datatype old_datatype, MPI_Datatype* new_datatype) -> int {
-	int old_datatype_size;  // NOLINT(cppcoreguidelines-init-variables)
-	MPI_Type_size(old_datatype, &old_datatype_size);
-
-	// return create_subarray_aux(lyt, 1, old_datatype, new_datatype);
-	skeleton const sk(lyt, old_datatype);
-	// new_datatype = std::move(sk).type();
-	{
-		MPI_Datatype vector_datatype;  // NOLINT(cppcoreguidelines-init-variables)
-		MPI_Type_create_hvector(
-			lyt.size(), 1,
-			lyt.stride() * old_datatype_size,
-			sk.type(), &vector_datatype
-		);
-
-		MPI_Type_create_resized(vector_datatype, 0, lyt.stride() * old_datatype_size, new_datatype);
-		MPI_Type_free(&vector_datatype);
-	}
-	return MPI_SUCCESS;
-}
-
-template<typename Size = int>
-class message : skeleton<Size> {
-	void* buf_;
-	// Size         count_;
-	// MPI_Datatype datatype_;
-
- public:
-	message(void* buf, skeleton<Size>&& sk) : skeleton<Size>{std::move(sk)}, buf_{buf} {}
-
-	template<class Layout>
-	message(void* buf, Layout const& lyt, MPI_Datatype dt) : skeleton<Size>(lyt, dt), buf_{buf} {}
-
-	template<class Arr>
-	explicit message(Arr const& arr)
-	: message{
-		  const_cast<void*>(static_cast<void const*>(arr.base())), skeleton<Size>{arr.layout(), MPI_INT}  // NOLINT(cppcoreguidelines-pro-type-const-cast)
-    } {}
-
-	message(message const& other) = delete;
-	message(message&&)            = delete;
-
-	auto operator=(message const&) = delete;
-	auto operator=(message&&)      = delete;
-
-	~message() = default;
-
-	auto buffer() const { return buf_; }
-	using skeleton<Size>::count;
-	// auto count() const { return this->count_; }
-	using skeleton<Size>::type;
-	// auto type() const { return this->datatype_; }
-
-	// template<std::size_t Index>
-	// std::tuple_element_t<Index, skeleton<>> const& get() const& {
-	//  if constexpr(Index == 0)
-	//      return buf_;
-	//  if constexpr(Index == 1)
-	//      return this->count_;
-	//  if constexpr(Index == 2)
-	//      return this->datatype_;
-	// }
-};
-
-}  // namespace boost::multi::mpi
-
-// namespace std {
-// template<> struct tuple_size<::boost::multi::mpi::skeleton<>> : integral_constant<size_t, 3> {};
-
-// template<> struct tuple_element<0, ::boost::multi::mpi::skeleton<>> {
-//  using type = void*;
-// };
-// template<> struct tuple_element<1, ::boost::multi::mpi::skeleton<>> {
-//  using type = int;
-// };
-// template<> struct tuple_element<2, ::boost::multi::mpi::skeleton<>> {
-//  using type = MPI_Datatype;
-// };
-// }  // namespace std
 
 namespace multi = boost::multi;
 
@@ -278,7 +62,7 @@ void test_1d(MPI_Comm comm) {
 		} else if(world_rank == 1) {
 			multi::array<int, 1> CC(3, 99);  // NOLINT(misc-const-correctness)
 
-			auto const C_msg = multi::mpi::message(CC);
+			auto const& C_msg = multi::mpi::message(CC.elements());
 
 			MPI_Recv(C_msg.buffer(), C_msg.count(), C_msg.type(), 0, 0, comm, MPI_STATUS_IGNORE);
 			BOOST_TEST(( CC == multi::array<double, 1>({1, 2, 3}) ));
@@ -343,7 +127,7 @@ void test_2d(MPI_Comm comm) {
 			auto const& BB = AA({0, 2}, {1, 3});
 			BOOST_TEST(( BB == multi::array<double, 2>({{2, 3}, {5, 6}}) ));
 
-			auto const B_msg = multi::mpi::message<>(BB);
+			auto const B_msg = multi::mpi::message(BB.elements());
 
 			MPI_Send(B_msg.buffer(), B_msg.count(), B_msg.type(), 1, 0, comm);
 		} else if(world_rank == 1) {
@@ -368,7 +152,7 @@ void test_2d(MPI_Comm comm) {
 			auto const& BB = AA({0, 2}, {1, 3});
 			BOOST_TEST(( BB == multi::array<double, 2>({{2, 3}, {5, 6}}) ));
 
-			auto const B_msg = multi::mpi::message<>(BB);
+			auto const B_msg = multi::mpi::message<>(BB.elements());
 
 			MPI_Send(B_msg.buffer(), B_msg.count(), B_msg.type(), 1, 0, comm);
 		} else if(world_rank == 1) {
@@ -409,6 +193,72 @@ void test_2d(MPI_Comm comm) {
 					  << CC[1][0] << ' ' << CC[1][1] << '\n';
 
 			BOOST_TEST(( CC == multi::array<double, 2>({{2, 3}, {5, 6}}) ));
+		}
+	}
+}
+
+void test_2d_int(MPI_Comm comm) {
+	int world_rank;  // NOLINT(cppcoreguidelines-init-variables)
+	MPI_Comm_rank(comm, &world_rank);
+	int world_size;  // NOLINT(cppcoreguidelines-init-variables)
+	MPI_Comm_size(comm, &world_size);
+
+	{
+		if(world_rank == 0) {
+			auto const AA = multi::array<int, 2>({
+				{1, 2, 3},
+				{4, 5, 6}
+			});
+
+			auto const& BB = AA({0, 2}, {1, 3});
+			BOOST_TEST(( BB == multi::array<int, 2>({{2, 3}, {5, 6}}) ));
+
+			auto const B_msg = multi::mpi::message(BB.elements());
+
+			MPI_Send(B_msg.buffer(), B_msg.count(), B_msg.type(), 1, 0, comm);
+		} else if(world_rank == 1) {
+			multi::array<int, 2> CC({2, 2}, 99);
+
+			auto const C_sk = multi::mpi::skeleton<int>(CC.layout());
+
+			MPI_Recv(CC.base(), C_sk.count(), C_sk.type(), 0, 0, comm, MPI_STATUS_IGNORE);
+			std::cout << CC[0][0] << ' ' << CC[0][1] << '\n'
+					  << CC[1][0] << ' ' << CC[1][1] << '\n';
+
+			BOOST_TEST(( CC == multi::array<int, 2>({{2, 3}, {5, 6}}) ));
+		}
+	}
+}
+
+void test_2d_double(MPI_Comm comm) {
+	int world_rank;  // NOLINT(cppcoreguidelines-init-variables)
+	MPI_Comm_rank(comm, &world_rank);
+	int world_size;  // NOLINT(cppcoreguidelines-init-variables)
+	MPI_Comm_size(comm, &world_size);
+
+	{
+		if(world_rank == 0) {
+			auto const AA = multi::array<double, 2>({
+				{1, 2, 3},
+				{4, 5, 6}
+			});
+
+			auto const& BB = AA({0, 2}, {1, 3});
+			BOOST_TEST(( BB == multi::array<double, 2>({{2, 3}, {5, 6}}) ));
+
+			auto const& B_msg = multi::mpi::message(BB.elements());
+
+			MPI_Send(B_msg.buffer(), B_msg.count(), B_msg.type(), 1, 0, comm);
+		} else if(world_rank == 1) {
+			multi::array<double, 2> CC({2, 2}, 99.0);
+
+			auto const C_sk = multi::mpi::skeleton<double>(CC.layout());
+
+			MPI_Recv(CC.base(), C_sk.count(), C_sk.type(), 0, 0, comm, MPI_STATUS_IGNORE);
+			std::cout << CC[0][0] << ' ' << CC[0][1] << '\n'
+					  << CC[1][0] << ' ' << CC[1][1] << '\n';
+
+			BOOST_TEST(( CC == multi::array<double, 2>({{2.0, 3.0}, {5.0, 6.0}}) ));
 		}
 	}
 }
@@ -466,6 +316,7 @@ auto main() -> int {  // NOLINT(bugprone-exception-escape)
 	MPI_Barrier(MPI_COMM_WORLD);
 
 	test_2d(MPI_COMM_WORLD);
+	test_2d_int(MPI_COMM_WORLD);
 
 	MPI_Finalize();
